+++ routers/goodses.js
@@ -18,7 +18,7 @@ const list = function* (next) {
   const {number, page} = this.params;
   const userList = yield goodsService.getGoodsList(number, page);
   if ("json" === this.way) {
-    this.set("Content-Type", "application/json");
+    this.set("Content-Type", "application/json; charset=utf-8");
     this.body = this.success(userList);
   }
   return yield next;